+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 21.0-SNAPSHOT - unreleased
 
+### ⚙️ Technical
+* Improvements to the ability to configure Hibernate Caches for 2nd-level and connection caches.
+
 ## 20.2.1 - 2024-07-09
 
 ### ⚙️ Technical